pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wholemem-mcp"
version = "0.1.0"
description = "Local-only privacy-respecting memory MCP server integrating Screenpipe, mem0, and Obsidian"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "mem0ai>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "watchdog>=4.0.0",
    "aiosqlite>=0.20.0",
    "zstandard>=0.22.0",
    "uvicorn>=0.30.0",
    "starlette>=0.37.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.35.0"]

[project.scripts]
wholemem-mcp = "wholemem_mcp.server:main"
wholemem-server = "wholemem_mcp.server:main"
wholemem-ui = "wholemem_mcp.ui.run:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wholemem_mcp"]

[tool.hatch.build.targets.sdist]
include = ["src/wholemem_mcp"]
